+++ packages/boethius-lang/src/builtins.js
@@ -13,9 +13,11 @@ import ScopeNode from './ScopeNode';
 import type { YY } from './types';
 import { Stringable } from './interfaces/Stringable';
 import { Serializable } from './interfaces/Serializable';
+import { Executable } from './interfaces/Executable';
+import { Node } from './interfaces/Node';
 
 const BUILTINS = {
-    csym: function (yy: YY, args: {}) {
+    csym (yy: YY, args: {}) {
         const value = args[0].serialize();
         const measure = args[1].value;
         const beat = args[2] ? args[2].value : 0;
@@ -23,7 +25,8 @@ const BUILTINS = {
 
         return chordSymbol;
     },
-    timesig: function (yy: YY, args: {}) {
+
+    timesig (yy: YY, args: {}) {
         const numerator = args[0].value;
         const denominator = args[1].value;
         const measure = args[2] ? args[2].value : 0;
@@ -38,7 +41,8 @@ const BUILTINS = {
 
         return timeSignature;
     },
-    layout: function (yy: YY, args: Array<Stringable>) {
+
+    layout (yy: YY, args: Array<Stringable>) {
         // args to layout must be even length
         if (args.length % 2 !== 0) {
             throw new Error('layout must have even number of arguments.');
@@ -50,21 +54,24 @@ const BUILTINS = {
 
         return yy.layout.set(props);
     },
-    page: function (yy: YY, args: {}) {
+
+    page (yy: YY, args: {}) {
         const page = new PageNode({
             systems: args[0].value
         });
 
         return page;
     },
-    system: function (yy: YY, args: Array<NumberNode>) {
+
+    system (yy: YY, args: Array<NumberNode>) {
         const [measures, ...lineSpacings] = args;
         const lineSpacing = lineSpacings.length ? lineSpacings.map(num => num.value) : [0];
         const system = new SystemNode({measures: measures.value, lineSpacing});
 
         return system;
     },
-    line: function (yy: YY, args: Array<Serializable>) {
+
+    line (yy: YY, args: Array<Serializable>) {
         const props = args.reduce(function (acc, arg) {
             if (arg instanceof Keyword) {
                 acc.voices.push(arg.serialize());
@@ -80,22 +87,39 @@ const BUILTINS = {
 
         return line;
     },
-    clef: function (yy: YY, args: Array<Serializable>) {
+
+    clef (yy: YY, args: Array<Serializable>) {
         const value = args[0].serialize();
         const measure = args[1] ? args[1].value : 0;
         const beat = args[2] ? args[2].value : 0;
         return new Clef({value, measure, beat});
     },
-    key: function (yy: YY, args: Array<Serializable>) {
+
+    key (yy: YY, args: Array<Serializable>) {
         const root = args[0].serialize();
         const mode = args[1].serialize();
         const measure = args[2] ? args[2].value : 0;
         const beat = args[3] ? args[3].value : 0;
         return new Key({root, mode, measure, beat});
     },
-    legato: function (yy: YY, args: Array<Serializable>) {
+
+    legato (yy: YY, args: Array<Serializable>) {
         return new ScopeNode({legato: uuid()}, args);
+    },
+
+    repeat (yy: YY, args: [NumberNode, Node]) {
+        const [repeats, ...items] = args;
+        let repeated: Array<Executable> = [];
+        for (let i = 0; i < repeats.value; i++) {
+            repeated = repeated.concat(items.map(clone));
+        }
+        return new ScopeNode({}, repeated);
     }
 };
+
+function clone (item: any) {
+    const Constructor = item.constructor;
+    return new Constructor(item.props, item.children);
+}
 
 export default BUILTINS;