+++ packages/boethius-lang/src/interfaces/Executable.js
@@ -1,5 +1,6 @@
 // @flow
 import type { YY } from '../types';
+import { Node } from './Node';
 
 export interface Executable {
     execute(yy: YY, scope: {}): {}