+++ packages/boethius-lang/src/interfaces/Node.js
@@ -1,0 +1,5 @@
+// @flow
+export interface Node {
+    props: any,
+    children?: Array<mixed>
+}